Add dek to StoryCard; separate StoryCard from VerticalCard; remove VerticalCard
