+++ components/StoryCard/StoryCard.styles.tsx
@@ -8,16 +8,16 @@ export const imageWrapper = 'transition-all aspect-w-1 aspect-h-1 overflow-hidde
 
 export const image = 'object-cover w-full h-full group-hocus-within:scale-105 transition-transform';
 
-export const heading = (hasTabColor: boolean) => cnb('mt-06em pr-08em pl-1em rs-mb-neg1 text-current', {
-  '@200:border-l-[1.2rem] @xs:border-l-[1.8rem] @200:pl-05em @200:pr-08em @320:pr-1em': hasTabColor,
+export const heading = (hasTabColor: boolean) => cnb('mt-06em pr-08em pl-21 rs-mb-neg1 text-current', {
+  '@200:border-l-[1.2rem] @xs:border-l-[1.8rem] @200:pl-12 @xs:pl-21 @200:pr-08em @320:pr-1em': hasTabColor,
 });
 
 export const headingLink = 'stretched-link no-underline !font-bold !leading-tight';
 
-export const ctaLink = 'inline-block w-fit stretched-link rs-mt-1 hocus:bg-gradient-to-l hocus:from-transparent hocus:to-digital-red';
-
-export const taxonomy = (hasTabColor: boolean) => cnb('list-unstyled leading-display children:mr-12 last:children:ml-0 mr-18', {
+export const taxonomy = (hasTabColor: boolean) => cnb('list-unstyled leading-display children:mr-12 last:children:ml-0 mr-18 ml-36', {
   '@200:ml-24 @xs:ml-36': hasTabColor,
 });
 
 export const taxonomyItem = 'inline-block mb-0';
+
+export const body = '@200:pl-12 @xs:pl-21 @200:pr-08em @320:pr-1em @200:ml-12 @xs:ml-18';
