+++ components/Storyblok/SbStoryCard.tsx
@@ -1,7 +1,7 @@
 import { storyblokEditable } from '@storyblok/react/rsc';
 import { type AnimationType } from '../Animate';
 import { type HeadingType } from '../Typography';
-import { StoryCard } from '../VerticalCard';
+import { StoryCard } from '../StoryCard';
 import { type SbImageType, type SbLinkType } from './Storyblok.types';
 import { paletteAccentColors, type PaletteAccentHexColorType } from '@/utilities/colorPalettePlugin';
 
@@ -12,6 +12,7 @@ export type SbStoryCardProps = {
       content?: {
         title?: string;
         cardTitle?: string;
+        dek?: string;
         topics?: string[];
         heroImage?: SbImageType;
         bgImage?: SbImageType;
@@ -38,6 +39,7 @@ export const SbStoryCard = ({
       content: {
         title = '',
         cardTitle = '',
+        dek = '',
         topics = [],
         heroImage: { filename: heroFilename = '', focus: heroFocus = '' } = {},
         bgImage: { filename: bgFilename = '', focus: bgFocus = '' } = {},
@@ -61,6 +63,7 @@ export const SbStoryCard = ({
     heading={heading || cardTitle || title}
     headingLevel={headingLevel}
     isSmallHeading={isSmallHeading}
+    body={dek}
     imageSrc={cardImage || mobileFilename || heroFilename || bgFilename }
     imageFocus={cardFocus || mobileFocus || heroFocus || bgFocus}
     tabColor={paletteAccentColors[value]}
