+++ components/ThemeCard/ThemeCard.styles.tsx
@@ -1,13 +1,11 @@
-export const root = '@container relative z-10 max-w-[32rem] sm:max-w-400 md:max-w-full mx-auto';
+export const root = 'relative z-10 max-w-[32rem] sm:max-w-400 md:max-w-full mx-auto';
 
-export const cardWrapper = 'relative group @200:text-15 @250:text-17 @280:!type-0 @md:!text-26';
+export const cardWrapper = 'group relative';
 
 export const imageWrapper = 'transition-all aspect-w-1 aspect-h-1 overflow-hidden';
 
 export const image = 'object-cover w-full h-full group-hocus-within:scale-105 transition-transform';
 
 export const heading = 'mt-06em rs-mb-neg1 text-current';
-
-export const headingLink = 'stretched-link no-underline !font-bold !leading-tight';
 
 export const ctaLink = 'inline-block w-fit stretched-link rs-mt-1 hocus:bg-gradient-to-l hocus:from-transparent hocus:to-digital-red';
