--- components/VerticalCard/index.ts
@@ -1,4 +1,0 @@
-export * from './StoryCard';
-export * from './ThemeCard';
-export * from './VerticalCard';
-export * from './VerticalCard.styles';
